4-29

+++ 食工119417刘源峰实验3-2.py
@@ -1,4 +1,4 @@
-x=int(eval(input("请输入x:")))
+x=float(eval(input("请输入x:")))
 if x<0 and x!=-3:
     ret=x**2+x-6
 elif 0<=x<10 and x!=2 and x!=3:
